Renomação de Lumen para Quintana
Finalizando a mudança de qualquer referência de Lumen para Quintana
